// src/models/AccountInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gerrit Code Review REST API
 * The Gerrit Code Review REST API. Paths, parameters, and payload schemas are generated statically from the server\'s REST bindings; human-authored summaries and descriptions are supplied here and merged in at build time.
 *
 * The version of the OpenAPI document: 3.15.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AvatarInfo } from './AvatarInfo';
import {
    AvatarInfoFromJSON,
    AvatarInfoFromJSONTyped,
    AvatarInfoToJSON,
    AvatarInfoToJSONTyped,
} from './AvatarInfo';

/**
 * 
 * @export
 * @interface AccountInfo
 */
export interface AccountInfo {
    /**
     * 
     */
    accountId?: number;
    /**
     * 
     */
    name?: string;
    /**
     * 
     */
    displayName?: string;
    /**
     * 
     */
    email?: string;
    /**
     * 
     */
    secondaryEmails?: Array<string>;
    /**
     * 
     */
    username?: string;
    /**
     * 
     */
    avatars?: Array<AvatarInfo>;
    /**
     * 
     */
    moreAccounts?: boolean;
    /**
     * 
     */
    status?: string;
    /**
     * 
     */
    inactive?: boolean;
    /**
     * 
     */
    deleted?: boolean;
    /**
     * 
     */
    tags?: Array<string>;
}

/**
 * Check if a given object implements the AccountInfo interface.
 */
export function instanceOfAccountInfo(value: object): value is AccountInfo {
    return true;
}

export function AccountInfoFromJSON(json: any): AccountInfo {
    return AccountInfoFromJSONTyped(json, false);
}

export function AccountInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccountInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'accountId': json['_account_id'] == null ? undefined : json['_account_id'],
        'name': json['name'] == null ? undefined : json['name'],
        'displayName': json['display_name'] == null ? undefined : json['display_name'],
        'email': json['email'] == null ? undefined : json['email'],
        'secondaryEmails': json['secondary_emails'] == null ? undefined : json['secondary_emails'],
        'username': json['username'] == null ? undefined : json['username'],
        'avatars': json['avatars'] == null ? undefined : ((json['avatars'] as Array<any>).map(AvatarInfoFromJSON)),
        'moreAccounts': json['_more_accounts'] == null ? undefined : json['_more_accounts'],
        'status': json['status'] == null ? undefined : json['status'],
        'inactive': json['inactive'] == null ? undefined : json['inactive'],
        'deleted': json['deleted'] == null ? undefined : json['deleted'],
        'tags': json['tags'] == null ? undefined : json['tags'],
    };
}

export function AccountInfoToJSON(json: any): AccountInfo {
    return AccountInfoToJSONTyped(json, false);
}

export function AccountInfoToJSONTyped(value?: AccountInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        '_account_id': value['accountId'],
        'name': value['name'],
        'display_name': value['displayName'],
        'email': value['email'],
        'secondary_emails': value['secondaryEmails'],
        'username': value['username'],
        'avatars': value['avatars'] == null ? undefined : ((value['avatars'] as Array<any>).map(AvatarInfoToJSON)),
        '_more_accounts': value['moreAccounts'],
        'status': value['status'],
        'inactive': value['inactive'],
        'deleted': value['deleted'],
        'tags': value['tags'],
    };
}